Use `size_t` for byte and element counts
Changes `bytes_len` and `elem_count` to `size_t` to match the return type of `bytes.size()` and handle potentially large sizes correctly.

PiperOrigin-RevId: 819560782

+++ tensorflow/compiler/mlir/lite/utils/const_tensor_utils.cc
@@ -67,10 +67,10 @@ template <typename T>
 llvm::SmallVector<mlir::APInt> ReadAsHostEndian(ArrayRef<uint8_t> bytes) {
   llvm::SmallVector<mlir::APInt> ret;
   size_t read_size = sizeof(T);
-  int bytes_len = bytes.size();
+  size_t bytes_len = bytes.size();
   assert(bytes_len % read_size == 0);
 
-  int elem_count = bytes_len / read_size;
+  size_t elem_count = bytes_len / read_size;
   ret.reserve(elem_count);
 
   const char* data_ptr = reinterpret_cast<const char*>(bytes.data());
